+++ frontend/src/domains/admin/Login/Logins.tsx
@@ -5,12 +5,12 @@ import FormField from '@/domains/admin/components/FormField/FormField';
 import {
   LoginField,
   loginFields,
-} from '@/domains/admin/Logins/constants/loginFields';
+} from '@/domains/admin/Login/constants/loginFields';
 import {
   fieldContainer,
   loginCaptionContainer,
   loginForm,
-} from '@/domains/admin/Logins/Login.style';
+} from '@/domains/admin/Login/Login.style';
 import useAuthForm from '@/domains/admin/hooks/useAuthForm';
 import { useAppTheme } from '@/hooks/useAppTheme';
 import {
